Additional Cosmetic Tweaks, To Do List
I added a couple of labels to the form to show created/last saved dates.
I also changed the font of the whole form to Comic Sans because I like
the handwritten feel.

I also added a To Do list where I'll be recording things I need to get
done, and ideas that I have that I'd like to implement.