Created NoteList class, Added Data Binding
I added a DataGridView Component that replaced the ListBox component for
showing the list of Notes. I also added DataBinding for this component.

In order to make this work, I had to create a NoteList class, which
implements the ITypedList interface. Implementing this interface allows
me to bind a list of objects to a control, but then still exercise fine
control over which parts of the object get displayed in the component.

Added event handler skeletons.

Added previewText member to Note class.