Added backup file, enable tabs in note text
I added a backup file that all notes are saved to when the program is
closed. I also catch the FileNotFound exception when loading notes so
that the user doesn't get a weird messaged when the start the program
for the first time ever.

I also enabled tabs in the text area, but I want to shorten them to four
spaces instead of eight.